requirements/common.txt
aldjemy==0.8.0
astroid==1.5.3
awesome-slugify==1.6.5
awscli==1.11.175
beautifulsoup4==4.6.0
billiard==3.5.0.3
bitly-api==0.3
bleach==2.1.1
boto==2.48.0
botocore==1.7.33
boto3==1.4.7
braintree==3.38.0
bugsnag==3.2.0
business-calendar==0.2.1
celery==4.1.0
cffi==1.11.2
channels==1.1.8
colorama==0.3.9
cryptography==2.1.1
cssselect==0.9.1
django_exceptional_middleware==1.2.1
djangorestframework==3.7.1
django-appconf==1.0.2
django-classy-tags==0.8.0
django-annoying==0.10.3
django-block-ip==0.2.0
django-celery==3.2.1
django-countries==5.0
django-compressor==2.2
django-crispy-forms==1.6.1
django-debug-toolbar==1.8
django-easy-maps==0.9.3
django-grappelli==2.10.1
django-geoposition==0.3.0
django-iban==0.3.1
django-libsass==0.7
django-leaflet==0.22.0
django-localflavor==1.5.2
django-mail-factory==0.20
django-location-field==2.0.3
django-markup-deprecated==0.0.3
django-memoize==2.1.0
django-mmc==1.6.5
django-modeltranslation==0.12.1
django-pandas==0.4.5
django-postman==3.5.1
django-redis-cache==1.7.1
django-secure==1.0.1
django-sql-explorer==1.1.1
django-statici18n==1.4.0
django-storages==1.6.5
django-taggit==0.22.1
django-transaction-hooks==0.2
django-treebeard==4.1.2
django-validate-model-attribute-assignment==2.0.2
Django==2.0
djrill==2.1.0
elasticsearch==5.4.0
elasticsearch-dsl==5.3.0
fuzzywuzzy==0.15.1
geopy==1.11.0
gdata==2.0.18
lazy-object-proxy==1.1
hashids==1.2.0
invoke==0.21.0
ipython==6.2.1
IPy==0.83
isort==4.2.15
Jinja2==2.9.6
jsonfield==2.0.2
libsass==0.13.3
linkedin==0.1.5
lockfile==0.12.2
mandrill==1.0.57
Markdown==2.6.9
markdown2==2.3.4
mccabe==0.6.1
neo4j-driver==1.5.0
newrelic==2.96.0.80
numpy==1.13.3
pandas==0.20.3
pep8==1.7.0
phonenumbers==8.8.4
Pillow==4.3.0
premailer==3.1.1
progressbar2==3.34.3
psycopg2==2.7.3.1
pyaml==17.10.0
pycountry==17.9.23
python-dateutil==2.6.1
python-Levenshtein==0.12.0
python-magic==0.4.13
python3-memcached==1.51
python-linkedin==4.1
PyTumblr==0.0.7
pytz==2017.2
rabbitmq==0.2.0
raven==6.2.1
requests==2.18.4
redis==2.10.6
reportlab==3.4.0
restless==2.1.1
simplejson==3.11.1
six==1.11.0
unicodecsv==0.14.1
urbanairship==3.0.1
urllib3==1.22
user-agents==1.1.0
uWSGI==2.0.15
vatnumber==1.2
Wand==0.4.4
wrapt==1.10.10
xhtml2pdf==0.0.6
